socket connection rework
